# backend/openapi.yaml
openapi: 3.0.0
info:
  title: Forge 软件发布管理平台 API
  description: 本文档提供了 Forge 软件发布管理平台的详细 API 文档，包括认证、软件管理、统计分析等各个模块的接口说明。
  version: 1.0.0
  contact:
    name: Forge 开发团队
    email: dev@forge.com
servers:
  - url: http://localhost:5000/api
    description: 开发服务器
  - url: https://api.forge.com/api
    description: 生产服务器
security:
  - BearerAuth: []
paths:
...
  /webhooks:
    get:
      tags:
        - Webhook
      summary: 获取Webhook日志列表
      description: 获取Webhook的调用日志列表
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          description: 页码
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          description: 每页数量
          schema:
            type: integer
            default: 20
        - name: space_id
          in: query
          description: 软件空间ID
          schema:
            type: integer
      responses:
        '200':
          description: 获取Webhook日志列表成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 获取Webhook日志列表成功
                  data:
                    type: object
                    properties:
                      webhooks:
                        type: array
                        description: Webhook日志列表
                        items:
                          $ref: '#/components/schemas/WebhookLog'
                      total:
                        type: integer
                        description: 总记录数
                        example: 100
                      pages:
                        type: integer
                        description: 总页数
                        example: 5
                      current_page:
                        type: integer
                        description: 当前页码
                        example: 1
                      per_page:
                        type: integer
                        description: 每页数量
                        example: 20
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /webhooks/{log_id}:
    get:
      tags:
        - Webhook
      summary: 获取Webhook日志详情
      description: 获取指定Webhook日志的详细信息
      security:
        - BearerAuth: []
      parameters:
        - name: log_id
          in: path
          required: true
          description: 日志ID
          schema:
            type: integer
      responses:
        '200':
          description: 获取Webhook日志详情成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 获取Webhook日志详情成功
                  data:
                    $ref: '#/components/schemas/WebhookLog'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /webhooks/retry/{log_id}:
    post:
      tags:
        - Webhook
      summary: 重试Webhook回调
      description: 重试失败的Webhook回调
      security:
        - BearerAuth: []
      parameters:
        - name: log_id
          in: path
          required: true
          description: 日志ID
          schema:
            type: integer
      responses:
        '200':
          description: 重试成功
          content:
            application/json:
              schema:
                $ref: '#/components/responses/Success'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /webhooks/test/{space_id}:
    post:
      tags:
        - Webhook
      summary: 测试Webhook回调
      description: 发送测试Webhook回调
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/SpaceId'
      responses:
        '200':
          description: 测试成功
          content:
            application/json:
              schema:
                $ref: '#/components/responses/Success'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer Token认证
  schemas:
...
    WebhookLog:
      type: object
      properties:
        id:
          type: integer
          description: 日志ID
          example: 1
        space_name:
          type: string
          description: 软件空间名称
          example: Forge
        event_type:
          type: string
          description: 事件类型
          example: download
        payload:
          type: string
          description: 请求数据
          example: '{"event": "download", "data": {...}}'
        response_status:
          type: integer
          description: 响应状态码
          example: 200
        response_body:
          type: string
          description: 响应体
          example: Success
        attempt_time:
          type: string
          format: date-time
          description: 尝试时间
          example: 2023-01-01T10:00:00Z
  parameters:
    SpaceId:
      name: space_id
      in: path
      required: true
      description: 软件空间ID
      schema:
        type: integer
...
  responses:
    Success:
      description: 操作成功
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: true
              message:
                type: string
                example: 操作成功
              data:
                type: object
                nullable: true
                example: null
    BadRequest:
      description: 请求参数错误
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              message:
                type: string
                example: 请求参数错误
              error:
                type: object
                properties:
                  code:
                    type: string
                    example: INVALID_PARAMETER
                  message:
                    type: string
                    example: 请求参数无效
    Unauthorized:
      description: 未授权
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              message:
                type: string
                example: 未授权，请登录
              error:
                type: object
                properties:
                  code:
                    type: string
                    example: UNAUTHORIZED
                  message:
                    type: string
                    example: 未授权访问
    Forbidden:
      description: 权限不足
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              message:
                type: string
                example: 禁止访问
              error:
                type: object
                properties:
                  code:
                    type: string
                    example: FORBIDDEN
                  message:
                    type: string
                    example: 权限不足
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              message:
                type: string
                example: 资源不存在
              error:
                type: object
                properties:
                  code:
                    type: string
                    example: RESOURCE_NOT_FOUND
                  message:
                    type: string
                    example: 资源不存在
    InternalError:
      description: 服务器内部错误
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              message:
                type: string
                example: 服务器内部错误
              error:
                type: object
                properties:
                  code:
                    type: string
                    example: INTERNAL_ERROR
                  message:
                    type: string
                    example: 服务器内部错误